+++ tests/test_domain/test_cpu_integration.py
@@ -3,16 +3,9 @@ import subprocess
 import pytest
 
 from chronus.SystemIntegration.cpu_info_service import LsCpuInfoService
-
+from tests.test_domain.fixtures import mock_subprocess_run
 
 # pytest fixture to mock the subprocess.run method
-@pytest.fixture
-def mock_subprocess_run(mocker):
-    return mocker.patch.object(
-        subprocess,
-        "run",
-        return_value=subprocess.CompletedProcess(args="lscpu", returncode=0, stdout=ls_cpu_output),
-    )
 
 
 @pytest.fixture
@@ -127,31 +120,3 @@ def xtest_get_frequency_when_cores_have_different_frequencies(mock_frequency_fil
 
     # Assert
     assert frequencies == expected_frequencies
-
-
-ls_cpu_output = """Architecture:        x86_64
-CPU op-mode(s):      32-bit, 64-bit
-Byte Order:          Little Endian
-CPU(s):              64
-On-line CPU(s) list: 0-63
-Thread(s) per core:  2
-Core(s) per socket:  32
-Socket(s):           1
-NUMA node(s):        1
-Vendor ID:           AuthenticAMD
-CPU family:          23
-Model:               49
-Model name:          AMD EPYC 7502P 32-Core Processor
-Stepping:            0
-CPU MHz:             2500.000
-CPU max MHz:         2500.0000
-CPU min MHz:         1500.0000
-BogoMIPS:            4990.86
-Virtualization:      AMD-V
-L1d cache:           32K
-L1i cache:           32K
-L2 cache:            512K
-L3 cache:            16384K
-NUMA node0 CPU(s):   0-63
-Flags:               fpu vme de pse tsc msr pae mce cx8 apic sep mtrr pge mca cmov pat pse36 clflush mmx fxsr sse sse2 ht syscall nx mmxext fxsr_opt pdpe1gb rdtscp lm constant_tsc rep_good nopl nonstop_tsc cpuid extd_apicid aperfmperf pni pclmulqdq monitor ssse3 fma cx16 sse4_1 sse4_2 movbe popcnt aes xsave avx f16c rdrand lahf_lm cmp_legacy svm extapic cr8_legacy abm sse4a misalignsse 3dnowprefetch osvw ibs skinit wdt tce topoext perfctr_core perfctr_nb bpext perfctr_llc mwaitx cpb cat_l3 cdp_l3 hw_pstate ssbd mba ibrs ibpb stibp vmmcall fsgsbase bmi1 avx2 smep bmi2 cqm rdt_a rdseed adx smap clflushopt clwb sha_ni xsaveopt xsavec xgetbv1 xsaves cqm_llc cqm_occup_llc cqm_mbm_total cqm_mbm_local clzero irperf xsaveerptr wbnoinvd arat npt lbrv svm_lock nrip_save tsc_scale vmcb_clean flushbyasid decodeassists pausefilter pfthreshold avic v_vmsave_vmload vgif v_spec_ctrl umip rdpid overflow_recov succor smca sme sev sev_es
-"""
